--- src/app/admin/(configuration)/modules/ui/ModuleModal.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const ModuleModal = () => {
-  return <div>ModuleModal</div>;
-};
