--- src/app/admin/(configuration)/modules/ui/SectionModal.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-export const SectionModal = () => {
-  return <div>SectionModal</div>;
-};
-
-export default SectionModal;
