+++ src/app/admin/(configuration)/modules/ui/index.ts
@@ -1,3 +1,4 @@
 export { FilterModules } from "./FilterModules";
 export { ModulesModalView } from "./ModulesModal";
 export { SkeletonModules } from "./SkeletonModules";
+export { DragSection } from "./DragModules";
